+++ tests/testthat/test-checks.R
@@ -179,6 +179,23 @@ test_that("workflow objects", {
   })
 })
 
+test_that("errors informatively when needed package isn't installed", {
+  # rstanarm is not installed during CI runs
+  # in contexts where it _is_ installed, skip the test.
+  skip_if(rlang::is_installed("rstanarm"))
+  stan_wflow <- workflow(mpg ~ ., parsnip::linear_reg(engine = "stan"))
+
+  expect_snapshot(
+    check_workflow(stan_wflow),
+    error = TRUE
+  )
+
+  expect_snapshot(
+    fit_resamples(stan_wflow, rsample::bootstraps(mtcars)),
+    error = TRUE
+  )
+})
+
 test_that("workflow objects (will not tune, tidymodels/tune#548)", {
   skip_if_not_installed("glmnet")